requirements.txt
openai>=1.50.0
mcp>=1.0.0
python-dotenv>=1.0.1
